Add Gate::authorize() to all Methods in Panel/FaqCourseController.php

+++ app/Http/Controllers/Panel/FaqCourseController.php
@@ -8,25 +8,30 @@ use App\Http\Requests\Panel\FaqCourse\UpdateFaqCourseRequest;
 use App\Models\Course;
 use App\Models\FaqCourse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class FaqCourseController extends Controller
 {
     public function index()
     {
+        Gate::authorize('view-faqs');
+
         $faqCourses = FaqCourse::orderByDesc('id')->paginate(5);
         return view('panel.faq_courses.index',compact('faqCourses'));
     }
 
     public function create()
     {
-        //        Gate::authorize('create-episode');
+        Gate::authorize('create-faq');
+
         $courses = Course::all();
         return view('panel.faq_courses.create',compact('courses'));
     }
 
     public function store(CreateFaqCourseRequest $request)
     {
-        //        Gate::authorize('store-episode');
+        Gate::authorize('store-faq');
+
         FaqCourse::create(
             $request->validated()
         );
@@ -36,13 +41,15 @@ class FaqCourseController extends Controller
 
     public function showFaqCourse(FaqCourse $faqCourse)
     {
+        Gate::authorize('view-faq-course');
+
         $course = $faqCourse->course()->first();
         return view('panel.faq_courses.faq_course',compact('course'));
     }
 
     public function edit(FaqCourse $faqCourse)
     {
-        //        Gate::authorize('edit-episodes');
+        Gate::authorize('edit-faqs');
 
         $courses = Course::all();
         return view('panel.faq_courses.edit',compact(['faqCourse','courses']));
@@ -50,7 +57,7 @@ class FaqCourseController extends Controller
 
     public function update(UpdateFaqCourseRequest $request, FaqCourse $faqCourse)
     {
-        //        Gate::authorize('update-episodes');
+        Gate::authorize('update-faqs');
 
         $faqCourse->update(
             $request->validated()
@@ -61,7 +68,7 @@ class FaqCourseController extends Controller
 
     public function destroy(FaqCourse $faqCourse, Request $request)
     {
-        //        Gate::authorize('delete-episodes');
+        Gate::authorize('delete-faqs');
 
         $faqCourse->delete();
         $request->session()->flash('status','سوال متداول مورد نظر با موفقیت حذف شد !');
